services/assistant-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
asyncpg==0.29.0
aiosqlite==0.19.0
httpx==0.25.0
aiohttp==3.9.0
openai==1.3.0
google-cloud-texttospeech==2.14.0
google-api-python-client==2.100.0
gtts==2.4.0
python-jose[cryptography]==3.3.0
requests==2.31.0
aiofiles==23.2.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
